requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
frozenlist==1.5.0
game_sdk==0.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.3.2
nest-asyncio==1.6.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.1
pydantic>=2.10.5
pydantic_core>=2.27.2
pyproject-api==1.8.0
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
telegram-plugin-gamesdk==0.1.0
tox==4.23.2
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.1
yarl==1.18.3